Add a publication for uploaded files list with search text
